refactor: merge +Reactions into +ScrollBehavior, collapse coordinator to 2-file split (#21627)
